Move AsGatewayConfigurationType attribute to PaymentBundle
