+++ src/Sylius/Bundle/PaymentBundle/Tests/DependencyInjection/SyliusPaymentExtensionTest.php
@@ -14,8 +14,10 @@ declare(strict_types=1);
 namespace Sylius\Bundle\PaymentBundle\Tests\DependencyInjection;
 
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionTestCase;
+use Sylius\Bundle\PaymentBundle\Attribute\AsGatewayConfigurationType;
 use Sylius\Bundle\PaymentBundle\Attribute\AsPaymentMethodsResolver;
 use Sylius\Bundle\PaymentBundle\DependencyInjection\SyliusPaymentExtension;
+use Sylius\Bundle\PaymentBundle\Tests\Stub\GatewayConfigurationTypeStub;
 use Sylius\Bundle\PaymentBundle\Tests\Stub\PaymentMethodsResolverStub;
 use Symfony\Component\DependencyInjection\Definition;
 
@@ -37,6 +39,26 @@ final class SyliusPaymentExtensionTest extends AbstractExtensionTestCase
         $this->assertContainerBuilderHasServiceDefinitionWithTag(
             'acme.payment_methods_resolver_with_attribute',
             AsPaymentMethodsResolver::SERVICE_TAG,
+            ['type' => 'test', 'label' => 'Test', 'priority' => 15],
+        );
+    }
+
+    /** @test */
+    public function it_autoconfigures_gateway_configuration_type_with_attribute(): void
+    {
+        $this->container->setDefinition(
+            'acme.gateway_configuration_type_with_attribute',
+            (new Definition())
+                ->setClass(GatewayConfigurationTypeStub::class)
+                ->setAutoconfigured(true),
+        );
+
+        $this->load();
+        $this->compile();
+
+        $this->assertContainerBuilderHasServiceDefinitionWithTag(
+            'acme.gateway_configuration_type_with_attribute',
+            AsGatewayConfigurationType::SERVICE_TAG,
             ['type' => 'test', 'label' => 'Test', 'priority' => 15],
         );
     }
