+++ src/Sylius/Bundle/PayumBundle/Tests/DependencyInjection/SyliusPayumExtensionTest.php
@@ -14,33 +14,10 @@ declare(strict_types=1);
 namespace Sylius\Bundle\PayumBundle\Tests\DependencyInjection;
 
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionTestCase;
-use Sylius\Bundle\PayumBundle\Attribute\AsGatewayConfigurationType;
 use Sylius\Bundle\PayumBundle\DependencyInjection\SyliusPayumExtension;
-use Sylius\Bundle\PayumBundle\Tests\Stub\GatewayConfigurationTypeStub;
-use Symfony\Component\DependencyInjection\Definition;
 
 final class SyliusPayumExtensionTest extends AbstractExtensionTestCase
 {
-    /** @test */
-    public function it_autoconfigures_gateway_configuration_type_with_attribute(): void
-    {
-        $this->container->setDefinition(
-            'acme.gateway_configuration_type_with_attribute',
-            (new Definition())
-                ->setClass(GatewayConfigurationTypeStub::class)
-                ->setAutoconfigured(true),
-        );
-
-        $this->load();
-        $this->compile();
-
-        $this->assertContainerBuilderHasServiceDefinitionWithTag(
-            'acme.gateway_configuration_type_with_attribute',
-            AsGatewayConfigurationType::SERVICE_TAG,
-            ['type' => 'test', 'label' => 'Test', 'priority' => 15],
-        );
-    }
-
     /** @test */
     public function it_loads_gateway_config_validation_groups_parameter_value_properly(): void
     {
